// src/lib/societa/types.ts
export type Societa = {
  id: number;
  slug: string;
  nome: string;
  fantallenatore: string;
  nicknameInstagram: string;
  squadraReale: string;
  stagioneIngresso: string;
  legaAttuale: string;
  girone: string;
  logo: string;
  ranking: number;
  puntiRanking: number;
  leader: boolean;
  badgeNewEntry: boolean;
  badgeNeopromossa: boolean;
  badgeCampioneSerieA: boolean;
};